+++ CHANGELOG.md
@@ -2,4 +2,9 @@
 
 ## 2012-08-02
 
-* Initial Relase
+* Initial Relase.
+
+## 2012-08-05
+
+* Changed map[string]interface{} to map[string] string.
+* Added single test.